more vc build patches from phrosy

+++ libs/net/net_main.c
@@ -481,7 +481,7 @@ NET_CheckNewConnections
 struct {
 	double      time;
 	int         op;
-	long        session;
+	intptr_t    session;
 } vcrConnect;
 
 qsocket_t  *
@@ -502,7 +502,7 @@ NET_CheckNewConnections (void)
 			if (recording) {
 				vcrConnect.time = host_time;
 				vcrConnect.op = VCR_OP_CONNECT;
-				vcrConnect.session = (long) ret;
+				vcrConnect.session = (intptr_t) ret;
 				Qwrite (vcrFile, &vcrConnect, sizeof (vcrConnect));
 				Qwrite (vcrFile, ret->address, NET_NAMELEN);
 			}
@@ -558,7 +558,7 @@ returns -1 if connection is invalid
 struct {
 	double      time;
 	int         op;
-	long        session;
+	intptr_t    session;
 	int         ret;
 	int         len;
 } vcrGetMessage;
@@ -602,7 +602,7 @@ NET_GetMessage (qsocket_t * sock)
 		if (recording) {
 			vcrGetMessage.time = host_time;
 			vcrGetMessage.op = VCR_OP_GETMESSAGE;
-			vcrGetMessage.session = (long) sock;
+			vcrGetMessage.session = (intptr_t) sock;
 			vcrGetMessage.ret = ret;
 			vcrGetMessage.len = _net_message_message.cursize;
 			Qwrite (vcrFile, &vcrGetMessage, 24);
@@ -613,7 +613,7 @@ NET_GetMessage (qsocket_t * sock)
 		if (recording) {
 			vcrGetMessage.time = host_time;
 			vcrGetMessage.op = VCR_OP_GETMESSAGE;
-			vcrGetMessage.session = (long) sock;
+			vcrGetMessage.session = (intptr_t) sock;
 			vcrGetMessage.ret = ret;
 			Qwrite (vcrFile, &vcrGetMessage, 20);
 		}
@@ -637,7 +637,7 @@ returns -1 if the connection died
 struct {
 	double      time;
 	int         op;
-	long        session;
+	intptr_t    session;
 	int         r;
 } vcrSendMessage;
 
@@ -662,7 +662,7 @@ NET_SendMessage (qsocket_t * sock, sizebuf_t *data)
 	if (recording) {
 		vcrSendMessage.time = host_time;
 		vcrSendMessage.op = VCR_OP_SENDMESSAGE;
-		vcrSendMessage.session = (long) sock;
+		vcrSendMessage.session = (intptr_t) sock;
 		vcrSendMessage.r = r;
 		Qwrite (vcrFile, &vcrSendMessage, 20);
 	}
@@ -692,7 +692,7 @@ NET_SendUnreliableMessage (qsocket_t * sock, sizebuf_t *data)
 	if (recording) {
 		vcrSendMessage.time = host_time;
 		vcrSendMessage.op = VCR_OP_SENDMESSAGE;
-		vcrSendMessage.session = (long) sock;
+		vcrSendMessage.session = (intptr_t) sock;
 		vcrSendMessage.r = r;
 		Qwrite (vcrFile, &vcrSendMessage, 20);
 	}
@@ -727,7 +727,7 @@ NET_CanSendMessage (qsocket_t * sock)
 	if (recording) {
 		vcrSendMessage.time = host_time;
 		vcrSendMessage.op = VCR_OP_CANSENDMESSAGE;
-		vcrSendMessage.session = (long) sock;
+		vcrSendMessage.session = (intptr_t) sock;
 		vcrSendMessage.r = r;
 		Qwrite (vcrFile, &vcrSendMessage, 20);
 	}
